requirements.txt
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.0
anthropic>=0.18.0
openai>=1.12.0
jinja2>=3.1.0
